+++ permissionsx/contrib/django/__init__.py
@@ -5,6 +5,3 @@ PermissionsX - Authorization for Django.
 :license:   BSD, see LICENSE for more details.
 
 """
-from __future__ import absolute_import
-
-from permissionsx.contrib.django.templatetags import register