+++ permissionsx/settings.py
@@ -5,9 +5,10 @@ PermissionsX - Authorization for Django.
 :license:   BSD, see LICENSE for more details.
 
 """
+from __future__ import absolute_import
+
 from django.conf import settings
 
 
 PERMISSIONSX_REDIRECT_URL = getattr(settings, 'PERMISSIONSX_REDIRECT_URL', settings.LOGIN_URL)
 PERMISSIONSX_LOGOUT_IF_DENIED = getattr(settings, 'PERMISSIONSX_LOGOUT_IF_DENIED', False)
-PERMISSIONSX_DEBUG = getattr(settings, 'PERMISSIONSX_DEBUG', False)