+++ runtests.py
@@ -6,6 +6,7 @@ PermissionsX - Authorization for Django.
 :license:   BSD, see LICENSE for more details.
 
 """
+from __future__ import absolute_import
 import sys
 import os